refactor: remove signature interface
