+++ vertx-web/src/test/java/io/vertx/ext/web/SignatureTest.java
@@ -1,5 +1,6 @@
 package io.vertx.ext.web;
 
+import io.vertx.ext.web.impl.Signature;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
@@ -8,14 +9,14 @@ public class SignatureTest {
 
   @Test
   public void verifiesItsOwnSignatures() {
-    Signature signature = Signature.create("any-string");
+    Signature signature = new Signature("any-string");
     final String token = signature.sign("my-random-bit-of-data");
     assertTrue(signature.verify(token));
   }
 
   @Test
   public void handlesMultiPeriodData() {
-    Signature signature = Signature.create("any-string");
+    Signature signature = new Signature("any-string");
     final String token = signature.sign("my-random-bit-of-data.some-other-data.bit");
     assertTrue(token.contains("my-random-bit-of-data.some-other-data.bit"));
     assertEquals(4, token.split("\\.").length);
@@ -24,13 +25,13 @@ public class SignatureTest {
 
   @Test
   public void rejectsDataWithoutSignature() {
-    Signature signature = Signature.create("any-string");
+    Signature signature = new Signature("any-string");
     assertFalse(signature.verify("just-some-data-not-signed"));
   }
 
   @Test
   public void rejectsAlteredData() {
-    Signature signature = Signature.create("any-string");
+    Signature signature = new Signature("any-string");
     String token = signature.sign("my-random-bit-of-data");
     token = token.replaceFirst("random-bit", "edited-bit");
     assertFalse(signature.verify(token));
@@ -38,14 +39,14 @@ public class SignatureTest {
 
   @Test
   public void parseSignedData() {
-    Signature signature = Signature.create("any-string");
+    Signature signature = new Signature("any-string");
     String token = signature.sign("my-random-bit-of-data");
     assertEquals("my-random-bit-of-data", signature.parse(token));
   }
 
   @Test
   public void parseAlteredSignedData() {
-    Signature signature = Signature.create("any-string");
+    Signature signature = new Signature("any-string");
     String token = signature.sign("my-random-bit-of-data");
     token = token.replaceFirst("random-bit", "edited-bit");
     assertNull(signature.parse(token));
